requirements.txt
inputs==0.5
pkg-resources==0.0.0
RPi.GPIO==0.6.5
